Setup variable speed for asynchronous scroll
Fixes #28

+++ src/matrix.rs
@@ -35,7 +35,7 @@ impl Matrix {
             }
             //  Generate the stream
             let height_offset = utils::random_between(-50, 0);
-            let stream = Stream::new(c as i32, height_offset, config);
+            let stream = Stream::new(c as f32, height_offset as f32, config);
 
             //  Add stream to vector collection
             ret.streams.push(stream);
